# ApiSpecyfication.json
{
  "swagger": "2.0",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "1.0.0"
  },
  "host": "www",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/isLoggedIn": {
      "get": {
        "tags": [
          "user",
          "session"
        ],
        "description": "Return true if user is logged into system",
        "responses": {
          "200": {
            "description": "Returns true if user is logged",
            "schema": {
              "type": "boolean"
            }
          },
          "400": {
            "description": "Request cannot be processed"
          }
        }
      }
    },
    "/getUserRole": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Returns user role",
        "responses": {
          "200": {
            "description": "Returns user role",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Request cannot be processed"
          }
        }
      }
    },
    "/getUserId": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Return user id",
        "responses": {
          "200": {
            "description": "Returns user role",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Request cannot be processed"
          }
        }
      }
    },
    "/getUser": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Return user",
        "responses": {
          "200": {
            "description": "Returns user",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Request cannot be processed"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "user": {
      "type": "object",
      "required": [
        "userId",
        "userName",
        "password",
        "firstName",
        "lastName",
        "email",
        "phoneNumber",
        "role"
      ],
      "properties": {
        "userId": {
          "type": "integer",
          "description": "Identifier in DB",
          "example": 12
        },
        "userName": {
          "type": "string",
          "description": "User nick name",
          "example": "Maciek13"
        },
        "password": {
          "type": "string",
          "description": "User password",
          "example": "idsahuioiuhfgr2390"
        },
        "firstName": {
          "type": "string",
          "description": "User first name",
          "example": "Marek"
        },
        "lastName": {
          "type": "string",
          "description": "User last name",
          "example": "Kowal"
        },
        "email": {
          "type": "string",
          "description": "User email",
          "example": "lubie@to.pl"
        },
        "phoneNumber": {
          "type": "string",
          "description": "User phone number",
          "example": "123456789"
        },
        "role": {
          "type": "object",
          "description": "Role of user in system",
          "$ref": "#/definitions/role"
        }
      }
    },
    "role": {
      "type": "object",
      "required": [
        "name",
        "roleId"
      ],
      "properties": {
        "roleId": {
          "type": "integer",
          "description": "Identifier in DB",
          "example": 13
        },
        "name": {
          "type": "string",
          "description": "Name of role",
          "example": "Admin"
        }
      }
    },
...
  }
}